+++ test/Integration/ApiGateway/GetAsyncShould.cs
@@ -1,9 +1,11 @@
 ﻿using System;
 using System.Net;
 using System.Net.Http;
+using System.Net.Http.Json;
 using System.Threading;
 using System.Threading.Tasks;
 using AwsSignatureVersion4.Integration.ApiGateway.Authentication;
+using AwsSignatureVersion4.Integration.ApiGateway.Contents;
 using AwsSignatureVersion4.Private;
 using Shouldly;
 using Xunit;
@@ -12,6 +14,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 {
     public class GetAsyncShould : ApiGatewayIntegrationBase
     {
+        private static readonly RichContent ExpectedResponseContent = new();
+
         public GetAsyncShould(IntegrationTestContext context)
             : base(context)
         {
@@ -33,6 +37,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         [Theory]
@@ -49,6 +55,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         #endregion
@@ -69,6 +77,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         [Theory]
@@ -85,6 +95,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         #endregion
@@ -109,6 +121,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         [Theory]
@@ -129,6 +143,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         #endregion
@@ -153,6 +169,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         [Theory]
@@ -173,6 +191,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         #endregion
@@ -197,6 +217,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         [Theory]
@@ -217,12 +239,14 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         [Theory]
         [InlineData(IamAuthenticationType.User)]
         [InlineData(IamAuthenticationType.Role)]
-        public void AbortGivenCanceled(IamAuthenticationType iamAuthenticationType)
+        public async Task AbortGivenCanceled(IamAuthenticationType iamAuthenticationType)
         {
             // Arrange
             var ct = new CancellationToken(true);
@@ -234,6 +258,11 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
                 Context.RegionName,
                 Context.ServiceName,
                 ResolveMutableCredentials(iamAuthenticationType));
+
+            while (task.Status == TaskStatus.WaitingForActivation)
+            {
+                await Task.Delay(1);
+            }
 
             // Assert
             task.Status.ShouldBe(TaskStatus.Canceled);
@@ -261,6 +290,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         [Theory]
@@ -281,6 +312,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         #endregion
@@ -307,6 +340,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         [Theory]
@@ -329,6 +364,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         #endregion
@@ -355,6 +392,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         [Theory]
@@ -377,6 +416,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         #endregion
@@ -401,6 +442,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         [Theory]
@@ -423,6 +466,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
 
         [Theory]
@@ -445,6 +490,8 @@ namespace AwsSignatureVersion4.Integration.ApiGateway
 
             // Assert
             response.StatusCode.ShouldBe(HttpStatusCode.OK);
+            var content = await response.Content.ReadFromJsonAsync<RichContent>();
+            content.ShouldBeEquivalentTo(ExpectedResponseContent);
         }
     }
 }